+++ src/AutoMapper/AutoMapper.csproj
@@ -162,22 +162,21 @@
     <Compile Include="QueryableExtensions\ExpressionRequest.cs" />
     <Compile Include="QueryableExtensions\ExpressionResolutionResult.cs" />
     <Compile Include="QueryableExtensions\Extensions.cs" />
-    <Compile Include="QueryableExtensions\IExpressionBinder.cs" />
+    <Compile Include="QueryableExtensions\IPropertyProjector.cs" />
     <Compile Include="QueryableExtensions\IExpressionResultConverter.cs" />
-    <Compile Include="QueryableExtensions\Impl\AssignableExpressionBinder.cs" />
-    <Compile Include="QueryableExtensions\Impl\CustomProjectionExpressionBinder.cs" />
-    <Compile Include="QueryableExtensions\Impl\EnumerableExpressionBinder.cs" />
-    <Compile Include="QueryableExtensions\Impl\MappedTypeExpressionBinder.cs" />
+    <Compile Include="QueryableExtensions\Impl\AssignablePropertyProjector.cs" />
+    <Compile Include="QueryableExtensions\Impl\EnumerablePropertyProjector.cs" />
+    <Compile Include="QueryableExtensions\Impl\MappedTypePropertyProjector.cs" />
     <Compile Include="QueryableExtensions\Impl\MemberGetterExpressionResultConverter.cs" />
     <Compile Include="QueryableExtensions\Impl\MemberResolverExpressionResultConverter.cs" />
-    <Compile Include="QueryableExtensions\Impl\NullableExpressionBinder.cs" />
+    <Compile Include="QueryableExtensions\Impl\NullablePropertyProjector.cs" />
     <Compile Include="QueryableExtensions\Impl\NullSubstitutionExpressionResultConverter.cs" />
     <Compile Include="QueryableExtensions\Impl\ParameterReplacementVisitor.cs" />
     <Compile Include="QueryableExtensions\Impl\QueryDataSourceInjection.cs" />
     <Compile Include="QueryableExtensions\Impl\QueryMapperVisitor.cs" />
     <Compile Include="QueryableExtensions\Impl\SourceInjectedQuery.cs" />
     <Compile Include="QueryableExtensions\Impl\SourceInjectedQueryProvider.cs" />
-    <Compile Include="QueryableExtensions\Impl\StringExpressionBinder.cs" />
+    <Compile Include="QueryableExtensions\Impl\ToStringPropertyProjector.cs" />
     <Compile Include="QueryableExtensions\IProjectionExpression.cs" />
     <Compile Include="QueryableExtensions\ProjectionExpression.cs" />
     <Compile Include="ResolutionContext.cs" />
