--- src/AutoMapper/QueryableExtensions/Impl/AssignableExpressionBinder.cs
@@ -1,21 +1,0 @@
-namespace AutoMapper.QueryableExtensions.Impl
-{
-    using System.Linq.Expressions;
-    using Internal;
-    using System.Reflection;
-
-    public class AssignablePropertyProjector : IPropertyProjector
-    {
-        public bool IsMatch(PropertyMap propertyMap, TypeMap propertyTypeMap, ExpressionResolutionResult result)
-        {
-            return propertyMap.DestinationPropertyType.IsAssignableFrom(result.Type);
-        }
-
-        public Expression Project(IMappingEngine mappingEngine, PropertyMap propertyMap, TypeMap propertyTypeMap,
-            ExpressionRequest request, ExpressionResolutionResult result,
-            IDictionary<ExpressionRequest, int> typePairCount)
-        {
-            return result.ResolutionExpression;
-        }
-    }
-}
