--- src/AutoMapper/QueryableExtensions/Impl/MappedTypeExpressionBinder.cs
@@ -1,19 +1,0 @@
-namespace AutoMapper.QueryableExtensions.Impl
-{
-    using System.Linq.Expressions;
-    using Internal;
-
-    public class MappedTypePropertyProjector : IPropertyProjector
-    {
-        public bool IsMatch(PropertyMap propertyMap, TypeMap propertyTypeMap, ExpressionResolutionResult result)
-        {
-            return propertyTypeMap != null;
-        }
-
-        public Expression Project(IMappingEngine mappingEngine, PropertyMap propertyMap, TypeMap propertyTypeMap,
-            ExpressionRequest request, ExpressionResolutionResult result, Internal.IDictionary<ExpressionRequest, int> typePairCount)
-        {
-            return ((IMappingEngineRunner)mappingEngine).CreateMapExpression(request, result.ResolutionExpression, typePairCount);
-        }
-    }
-}
