+++ src/AutoMapper/QueryableExtensions/Impl/NullablePropertyProjector.cs
@@ -1,0 +1,45 @@
+namespace AutoMapper.QueryableExtensions.Impl
+{
+    using System.Linq.Expressions;
+    using Internal;
+
+    public class NullablePropertyProjector : IPropertyProjector
+    {
+        public bool IsMatch(PropertyMap propertyMap, TypeMap propertyTypeMap, ExpressionResolutionResult result)
+        {
+            return propertyMap.DestinationPropertyType.IsNullableType()
+                   && !result.Type.IsNullableType();
+        }
+
+        public Expression Project(IMappingEngine mappingEngine, PropertyMap propertyMap, TypeMap propertyTypeMap,
+            ExpressionRequest request, ExpressionResolutionResult result,
+            IDictionary<ExpressionRequest, int> typePairCount)
+        {
+            if (result.ResolutionExpression.NodeType == ExpressionType.MemberAccess)
+            {
+                var memberExpr = (MemberExpression) result.ResolutionExpression;
+                if (memberExpr.Expression != null && memberExpr.Expression.NodeType == ExpressionType.MemberAccess)
+                {
+                    var destType = propertyMap.DestinationPropertyType;
+                    var parentExpr = memberExpr.Expression;
+                    Expression expressionToBind = Expression.Convert(memberExpr, destType);
+                    var nullExpression = Expression.Convert(Expression.Constant(null), destType);
+                    while (parentExpr.NodeType != ExpressionType.Parameter)
+                    {
+                        memberExpr = (MemberExpression) memberExpr.Expression;
+                        parentExpr = memberExpr.Expression;
+                        expressionToBind = Expression.Condition(
+                            Expression.Equal(memberExpr, Expression.Constant(null)),
+                            nullExpression,
+                            expressionToBind
+                            );
+                    }
+
+                    return expressionToBind;
+                }
+            }
+
+            return Expression.Convert(result.ResolutionExpression, propertyMap.DestinationPropertyType);
+        }
+    }
+}
