--- src/AutoMapper/QueryableExtensions/Impl/StringExpressionBinder.cs
@@ -1,18 +1,0 @@
-namespace AutoMapper.QueryableExtensions.Impl
-{
-    using System.Linq.Expressions;
-    using Internal;
-
-    public class ToStringPropertyProjector : IPropertyProjector
-    {
-        public bool IsMatch(PropertyMap propertyMap, TypeMap propertyTypeMap, ExpressionResolutionResult result)
-        {
-            return propertyMap.DestinationPropertyType == typeof(string);
-        }
-
-        public Expression Project(IMappingEngine mappingEngine, PropertyMap propertyMap, TypeMap propertyTypeMap, ExpressionRequest request, ExpressionResolutionResult result, IDictionary<ExpressionRequest, int> typePairCount)
-        {
-            return Expression.Call(result.ResolutionExpression, "ToString", null, null);
-        }
-    }
-}
